break out type-utils.ts from metamodel.ts (using an agent)
